@@ -4,7 +4,7 @@ version: 5.2.2
 homepage: https://github.com/greensopinion/dart-vector-tile-renderer
 
 environment:
-  sdk: ">=2.17.0 <4.0.0"
+  sdk: ">=3.0.0 <4.0.0"
   flutter: '>=3.2.0'
 
 dependencies:
